+++ man-roxygen/param_xss.R
@@ -1,0 +1,3 @@
+#' @param xss (list of named `list()`)\cr
+#' Lists of arguments for the function e.g. `list(list(x1 = 1, x2 = 2), list(x1 = 3, x2 = 4))`.
+#' If `xss` is empty, no tasks are created and the method returns an empty `character()`.