+++ man-roxygen/param_yss.R
@@ -1,0 +1,2 @@
+#' @param yss (list of named `list()`)\cr
+#' Lists of results for the function e.g. `list(list(y1 = 1, y2 = 2), list(y1 = 3, y2 = 4))`.