+++ man-roxygen/param_yss_extra.R
@@ -1,0 +1,3 @@
+#' @param yss_extra (list of named `list()`)\cr
+#' List of additional information stored along with the results
+#' e.g. `list(list(timestamp_ys = Sys.time()), list(timestamp_ys = Sys.time()))`.